Separate unecessary calls from logger
